+++ lib/facebook.js
@@ -16,8 +16,6 @@ function sendTextMessage(sender, message, callback) {
 }
 
 function sendMessage(sender, message, callback) {
-  const json = JSON.stringify(message)
-
   Request({
     url: FACEBOOK_URL + "/me/messages",
     qs: { access_token: ACCESS_TOKEN },
