+++ lib/user-store.js
@@ -8,14 +8,10 @@ AWS.config.update({
   region: 'eu-west-1'
 })
 
-module.exports = UserStore
+const docClient = new AWS.DynamoDB.DocumentClient()
 
-function UserStore() {
-  this._docClient = new AWS.DynamoDB.DocumentClient()
-}
-
-UserStore.prototype.getUser = function(id, callback) {
-  this._docClient.get({
+exports.getUser = (id, callback) => {
+  docClient.get({
     TableName: TABLE_NAME,
     Key: {
       "ID": id
@@ -23,8 +19,8 @@ UserStore.prototype.getUser = function(id, callback) {
   }, callback)
 }
 
-UserStore.prototype.addUser = function(id, front, callback) {
-  this._docClient.put({
+exports.addUser = (id, front, callback) => {
+  docClient.put({
     TableName: TABLE_NAME,
     Item: {
       "ID": id,
@@ -36,8 +32,8 @@ UserStore.prototype.addUser = function(id, front, callback) {
   }, callback)
 }
 
-UserStore.prototype.setNotificationTime = function(id, offset, time, timeUTC, callback) {
-  this._docClient.update({
+exports.setNotificationTime = (id, offset, time, timeUTC, callback) => {
+  docClient.update({
     TableName: TABLE_NAME,
     Key: {
       "ID": id
@@ -51,8 +47,8 @@ UserStore.prototype.setNotificationTime = function(id, offset, time, timeUTC, ca
   }, callback)
 }
 
-UserStore.prototype.setFront = function(id, front, callback) {
-  this._docClient.update({
+exports.setFront = (id, front, callback) => {
+  docClient.update({
     TableName: TABLE_NAME,
     Key: {
       "ID": id
@@ -64,8 +60,8 @@ UserStore.prototype.setFront = function(id, front, callback) {
   }, callback)
 }
 
-UserStore.prototype.notifyUsers = function(time, callback) {
-  this._docClient.query({
+exports.getUsersByTime = (time, callback) => {
+  docClient.query({
     TableName: TABLE_NAME,
     IndexName: "notificationTimeUTC-ID-index",
     KeyConditionExpression: "notificationTimeUTC = :t",
@@ -75,8 +71,8 @@ UserStore.prototype.notifyUsers = function(time, callback) {
   }, callback)
 }
 
-UserStore.prototype.unsubscribe = function(id, callback) {
-  this._docClient.update({
+exports.unsubscribe = (id, callback) => {
+  docClient.update({
     TableName: TABLE_NAME,
     Key: {
       "ID": id
@@ -88,4 +84,3 @@ UserStore.prototype.unsubscribe = function(id, callback) {
     }
   }, callback)
 }
-
